Feature(cubejs-query-orchestrator): Move source to src/ folder
